rename(#95): network를 data/network로 옮김
