+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/request/UpdateMindMap.kt
@@ -1,4 +1,4 @@
-package boostcamp.and07.mindsync.network.request
+package boostcamp.and07.mindsync.data.network.request
 
 import boostcamp.and07.mindsync.data.crdt.SerializedOperation
 import kotlinx.serialization.Serializable
