+++ AOS/app/src/main/java/boostcamp/and07/mindsync/data/network/response/OperationFromServer.kt
@@ -1,4 +1,4 @@
-package boostcamp.and07.mindsync.network.response
+package boostcamp.and07.mindsync.data.network.response
 
 import boostcamp.and07.mindsync.data.crdt.SerializedOperation
 import kotlinx.serialization.Serializable
