+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/MindSyncApplication.kt
